Correction du problème de connexion
