+++ login.html
@@ -65,6 +65,6 @@
   </footer>
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-  <script type="module" src="/src/login.js"></script>
+  <script type="module" src="/src/main.js"></script>
 </body>
 </html>
