+++ signin.html
@@ -63,6 +63,6 @@
   </div>
 
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-  <script type="module" src="/src/inscription.js"></script>
+  <script type="module" src="/src/main.js"></script>
 </body>
 </html>
