+++ src/main.js
@@ -1,2 +1,3 @@
 import '../node_modules/bootstrap/dist/js/bootstrap.bundle.min.js';
-import './inscription.js';
+import './login.js';
+import './signin.js';
